requirement.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==3.0.10
Flask-Script==2.0.6
flatbuffers==2.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
onnxruntime==1.11.0
opencv-python==4.5.5.64
Pillow==9.1.0
protobuf==3.20.0
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
watchdog==2.1.9
Werkzeug==2.1.1
wincertstore==0.2
zipp==3.8.0